requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
# Django==4.1.3
Django==3.2.16
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
idna==3.4
line-bot-sdk==2.3.0
multidict==6.0.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
yarl==1.8.1
